first attempt at adding Designers images
kept getting 404 errors for any other file I tried to import into the .pug files, so instead I hosted the files via telsa, which worked. Not sure if we will be able to keep the shape the designers wanted for the text boxes but I'm going to continue trying, this is far from complete, this is merely a save commit.